Refactor UI into modular components
