+++ src/ui/journal_window.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "journal.h"
+
+void DrawJournalWindow(Journal::Journal& journal);
+
